requirements.txt
insightface
onnxruntime
numpy
pillow
scikit-learn
opencv-python
